Major revision of sample preparation pipeline
