+++ .gitignore
@@ -10,3 +10,6 @@
 .DS_Store
 */.DS_Store
 *.json
+/SamplePrep/Outputs
+*/*.pyc
+*/*/*.pyc
