+++ SamplePrep/ROOTToH5/make_all.sh
@@ -1,31 +1,11 @@
-mkdir Outputs
+#!/bin/bash
+# source make_all.sh {ROOT_path} {sample_path}
 
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000001.pool.root.1
-mv output.h5 Outputs/output_1.h5
+mkdir -p $2/H5
 
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000002.pool.root.1
-mv output.h5 Outputs/output_2.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000003.pool.root.1
-mv output.h5 Outputs/output_3.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000004.pool.root.1
-mv output.h5 Outputs/output_4.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000005.pool.root.1
-mv output.h5 Outputs/output_5.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000006.pool.root.1
-mv output.h5 Outputs/output_6.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000007.pool.root.1
-mv output.h5 Outputs/output_7.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000008.pool.root.1
-mv output.h5 Outputs/output_8.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000009.pool.root.1
-mv output.h5 Outputs/output_9.h5
-
-./build/x86_64-slc6-gcc62-opt/bin/dump-xaod /eos/user/m/mazhang/LepIso/MC/MUON5/mc16_13TeV.410470.PhPy8EG_A14_ttbar_hdamp258p75_nonallhad.deriv.DAOD_MUON5.e6337_e5984_s3126_r10201_r10210_p3584/DAOD_MUON5.14537525._000010.pool.root.1
-mv output.h5 Outputs/output_10.h5
+for i in {1..30}
+do
+    printf -v padded_i "%06d" ${i}
+    ./build/x86_64-slc6-gcc62-opt/bin/dump-xaod $1/DAOD_MUON5.14537525._${padded_i}.pool.root.1
+    mv output.h5 $2/H5/output_${i}.h5
+done
